+++ packages/hasura-auth-js/src/types/hasura-auth/responses.ts
@@ -58,6 +58,12 @@ export type SignInPATResponse = NhostSessionResponse
 /** payload from hasura-auth endpoint /signin/anonymous */
 export type PasswordlessSmsOtpResponse = NhostSessionResponse
 
+/** payload from hasura-auth endpoint /signin/otp/email */
+export interface SignInEmailOTPResponse extends NullableErrorResponse {}
+
+/** payload from hasura-auth endpoint /signin/otp/email/verify */
+export type VerifyEmailOTPResponse = NhostSessionResponse
+
 /** payload from hasura-auth endpoint /signin/mfa/totp */
 export type SignInMfaTotpResponse = NhostSessionResponse