+++ packages/hasura-auth-js/src/types/options.ts
@@ -35,6 +35,8 @@ export interface RedirectOption {
 
 export interface PasswordlessOptions extends RegistrationOptions, RedirectOption {}
 export interface SignUpOptions extends RegistrationOptions, RedirectOption {}
+export interface EmailOTPOptions extends RegistrationOptions, RedirectOption {}
+
 export interface SignUpSecurityKeyOptions extends SignUpOptions {
   /** Optional nickname for the security key */
   nickname?: string