+++ examples/codegen-react-apollo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nhost-examples/codegen-react-apollo
 
+## 0.4.14
+
+### Patch Changes
+
+- Updated dependencies [fe6e8e2]
+  - @nhost/react@3.7.0
+  - @nhost/react-apollo@14.0.0
+
 ## 0.4.13
 
 ### Patch Changes