+++ examples/nextjs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nhost-examples/nextjs
 
+## 0.3.14
+
+### Patch Changes
+
+- Updated dependencies [fe6e8e2]
+  - @nhost/react@3.7.0
+  - @nhost/react-apollo@14.0.0
+  - @nhost/nextjs@2.1.23
+
 ## 0.3.13
 
 ### Patch Changes