+++ integrations/react-urql/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nhost/react-urql
 
+## 11.0.0
+
+### Patch Changes
+
+- Updated dependencies [fe6e8e2]
+  - @nhost/react@3.7.0
+
 ## 10.0.0
 
 ### Patch Changes