+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nhost/nextjs
 
+## 2.1.23
+
+### Patch Changes
+
+- Updated dependencies [fe6e8e2]
+  - @nhost/react@3.7.0
+
 ## 2.1.22
 
 ### Patch Changes